Update nextjs to have most logic in own component, not in pages
